Cleanup branch testing correct user for upload
