+++ MetaExchangeSowaLabs.Tests/MetaExchangeBestPriceTests.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using MetaExchangeSowaLabs.CustomErrors;
 using MetaExchangeSowaLabs.Enums;
+using MetaExchangeSowaLabs.Services;
+using MetaExchangeSowaLabs.Services.Interfaces;
 using Newtonsoft.Json;
 using Xunit;
 
@@ -9,44 +11,57 @@ namespace MetaExchangeSowaLabs.Tests
 {
     public class MetaExchangeBestPriceTests
     {
-
         [Fact]
         public void SendingEmptyListOrderBooksShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>();
-            var orderBooksUserBalance = new List<string>{@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
+            var orderBooksUserBalance = new List<string> {@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 15;
+
             
-            Assert.Throws<EmptyOrderBookOrBalanceException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+            Assert.Throws<EmptyOrderBookOrBalanceException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder,
+                    amountOfBtc));
         }
-        
+
         [Fact]
         public void SendingEmptyListOrderBooksUserBalanceShouldThrowException()
         {
-            var orderBooks = new List<string>{@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
+            var orderBooks = new List<string> {@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
             var orderBooksUserBalance = new List<string>();
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 15;
-        
-            Assert.Throws<EmptyOrderBookOrBalanceException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<EmptyOrderBookOrBalanceException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder,
+                    amountOfBtc));
         }
-        
+
         [Theory]
         [InlineData(0)]
         [InlineData(-2)]
         public void SendingNegativeOrZeroBtcAmountShouldThrowException(decimal amountOfBtc)
         {
-            var orderBooks = new List<string>{@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
-            var orderBooksUserBalance = new List<string>{@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
+            var orderBooks = new List<string> {@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
+            var orderBooksUserBalance = new List<string> {@"1548759600.25189	{""Bitcoin"": 5, ""Eur"": 1221312301.12}"};
             var typeOfOrder = TypeOfOrderEnum.Buy;
 
-            Assert.Throws<IllegalAmountOfBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+            Assert.Throws<IllegalAmountOfBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void BadlyFormatedJsonShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189
@@ -65,13 +80,16 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 15;
-        
-            Assert.Throws<JsonException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<JsonException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void UserWithNoMoneyWhenBuyingBtcShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -82,7 +100,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -90,13 +108,16 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 15;
-        
-            Assert.Throws<UserHasNoMoneyOrBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<UserHasNoMoneyOrBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void UserWithNoBtcWhenSellingBtcShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -107,7 +128,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -115,31 +136,38 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Sell;
             var amountOfBtc = (decimal) 15;
-        
-            Assert.Throws<UserHasNoMoneyOrBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<UserHasNoMoneyOrBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
-        
+
+
         [Theory]
         [MemberData(nameof(TestData))]
-        public void BuyingOrSellingBitcoinsShouldProvideCorrectStepsTheory(Dictionary<string, List<string>> expectedResult, IEnumerable<string> orderBooks, IEnumerable<string> orderBooksUserBalance,
+        public void BuyingOrSellingBitcoinsShouldProvideCorrectStepsTheory(
+            Dictionary<string, List<string>> expectedResult, IEnumerable<string> orderBooks,
+            IEnumerable<string> orderBooksUserBalance,
             TypeOfOrderEnum typeOfOrder, decimal amountOfBtc)
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
             
-            Assert.Equal(expectedResult,Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+            Assert.Equal(expectedResult,
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         public static IEnumerable<object[]> TestData()
         {
             yield return new object[]
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Buy 0.405 BTC at 2980.61 EUR",
-                        "Buy 0.405 BTC at 2980.62 EUR"
-                    }}
+                        "1548759600.25189", new List<string>
+                        {
+                            "Buy 0.405 BTC at 2980.61 EUR",
+                            "Buy 0.405 BTC at 2980.62 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -151,25 +179,27 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
                 },
                 new List<string>
                 {
-                   @"1548759600.25189	{""Bitcoin"": 1, ""Eur"": 12301.12}",
+                    @"1548759600.25189	{""Bitcoin"": 1, ""Eur"": 12301.12}",
                 },
                 TypeOfOrderEnum.Buy,
                 (decimal) 0.810
             };
-            
+
             yield return new object[]
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Sell 0.23 BTC at 2960.64 EUR",
-                        "Sell 0.2323 BTC at 2960.63 EUR"
-                    }}
+                        "1548759600.25189", new List<string>
+                        {
+                            "Sell 0.23 BTC at 2960.64 EUR",
+                            "Sell 0.2323 BTC at 2960.63 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -181,7 +211,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Buy"",""Kind"":""Limit"",""Amount"":0.2323,""Price"":2960.61}}],
                     ""Asks"":[{""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.64}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}}] 
-                    }", 
+                    }",
                 },
                 new List<string>
                 {
@@ -190,17 +220,19 @@ namespace MetaExchangeSowaLabs.Tests
                 TypeOfOrderEnum.Sell,
                 (decimal) 0.4623
             };
-            
+
             yield return new object[]
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Sell 0.23 BTC at 2960.64 EUR",
-                        "Sell 0.2323 BTC at 2960.63 EUR",
-                        "Sell 0.23 BTC at 2960.62 EUR"
-                    }}
+                        "1548759600.25189", new List<string>
+                        {
+                            "Sell 0.23 BTC at 2960.64 EUR",
+                            "Sell 0.2323 BTC at 2960.63 EUR",
+                            "Sell 0.23 BTC at 2960.62 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -234,15 +266,19 @@ namespace MetaExchangeSowaLabs.Tests
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Sell 0.23 BTC at 2960.64 EUR",
-                        "Sell 0.2323 BTC at 2960.63 EUR"
-                    }},
-                    {"1548759600.25190", new List<string>
+                        "1548759600.25189", new List<string>
+                        {
+                            "Sell 0.23 BTC at 2960.64 EUR",
+                            "Sell 0.2323 BTC at 2960.63 EUR"
+                        }
+                    },
                     {
-                        "Sell 0.23 BTC at 2960.64 EUR"
-                    }}
+                        "1548759600.25190", new List<string>
+                        {
+                            "Sell 0.23 BTC at 2960.64 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -273,17 +309,19 @@ namespace MetaExchangeSowaLabs.Tests
                 TypeOfOrderEnum.Sell,
                 (decimal) 0.6923
             };
-            
+
             yield return new object[]
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Buy 0.2323 BTC at 2960.61 EUR",
-                        "Buy 0.23 BTC at 2960.62 EUR",
-                        "Buy 0.23 BTC at 2960.63 EUR"
-                    }}
+                        "1548759600.25189", new List<string>
+                        {
+                            "Buy 0.2323 BTC at 2960.61 EUR",
+                            "Buy 0.23 BTC at 2960.62 EUR",
+                            "Buy 0.23 BTC at 2960.63 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -313,20 +351,24 @@ namespace MetaExchangeSowaLabs.Tests
                 TypeOfOrderEnum.Buy,
                 (decimal) 0.6923
             };
-            
-             yield return new object[]
+
+            yield return new object[]
             {
                 new Dictionary<string, List<string>>
                 {
-                    {"1548759600.25189", new List<string>
                     {
-                        "Buy 0.2323 BTC at 2960.61 EUR",
-                        "Buy 0.2277 BTC at 2960.62 EUR",
-                    }},
-                    {"1548759600.25190", new List<string>
+                        "1548759600.25189", new List<string>
+                        {
+                            "Buy 0.2323 BTC at 2960.61 EUR",
+                            "Buy 0.2277 BTC at 2960.62 EUR",
+                        }
+                    },
                     {
-                        "Buy 0.2323 BTC at 2960.61 EUR"
-                    }}
+                        "1548759600.25190", new List<string>
+                        {
+                            "Buy 0.2323 BTC at 2960.61 EUR"
+                        }
+                    }
                 },
                 new List<string>
                 {
@@ -358,11 +400,13 @@ namespace MetaExchangeSowaLabs.Tests
                 (decimal) 0.6923
             };
         }
-        
-        
+
+
         [Fact]
         public void UserWithNotEnoughBtcWhenSellingBtcShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -373,7 +417,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -381,13 +425,16 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Sell;
             var amountOfBtc = (decimal) 14.222;
-        
-            Assert.Throws<UserHasNoMoneyOrBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<UserHasNoMoneyOrBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void UserWithNotEnoughMoneyWhenBuyingBtcShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -398,7 +445,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -406,13 +453,16 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 0.5;
-        
-            Assert.Throws<CantBuyDesiredBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<CantBuyDesiredBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void UserWantsToBuyMoreThanAvailableBtcInTheMarketShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -423,7 +473,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -431,13 +481,16 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Buy;
             var amountOfBtc = (decimal) 100;
-        
-            Assert.Throws<CantBuyDesiredBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<CantBuyDesiredBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
+
         [Fact]
         public void UserWantsToSellMoreBtcThanOthersWantToBuyInTheMarketShouldThrowException()
         {
+            IMetaExchangeService metaExchangeService = new MetaExchangeService();
+            
             var orderBooks = new List<string>
             {
                 @"1548759600.25189	{
@@ -448,7 +501,7 @@ namespace MetaExchangeSowaLabs.Tests
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.63}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.62}},
                               {""Order"":{""Id"":null,""Time"":""0001-01-01T00:00:00"",""Type"":""Sell"",""Kind"":""Limit"",""Amount"":0.405,""Price"":2980.61}}] 
-                    }", 
+                    }",
             };
             var orderBooksUserBalance = new List<string>
             {
@@ -456,9 +509,9 @@ namespace MetaExchangeSowaLabs.Tests
             };
             var typeOfOrder = TypeOfOrderEnum.Sell;
             var amountOfBtc = (decimal) 100;
-        
-            Assert.Throws<CantSellDesiredBtcException>(() => Program.MetaExchangeBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
+
+            Assert.Throws<CantSellDesiredBtcException>(() =>
+                metaExchangeService.GetOptimalStepsForBestPrice(orderBooks, orderBooksUserBalance, typeOfOrder, amountOfBtc));
         }
-        
     }
 }
